Documentation Added + minor improvements
commit b9d94ab0789cdd292fe40c1edf68a102dfad9923
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 12 10:28:12 2025 +0200

    Docs Updated

commit 6bf09dfc4f34c10fde21b78e1d1ad07763f42994
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Fri Sep 12 00:55:21 2025 +0200

    HAWKI Username as Parameter

commit 584c8aeb5fb57b933b69c6096910884d6632ecf5
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Thu Sep 11 21:06:40 2025 +0200

    Bugfix: download file problem solved

commit 8eb521446d770b94d278d62f5df090c429a7681f
Author: Ariansdf <38947447+Ariansdf@users.noreply.github.com>
Date:   Thu Sep 11 18:43:18 2025 +0200

    imported documentation

+++ config/filesystems.php
@@ -85,7 +85,7 @@ return [
         'sftp' => [
             'driver' => 'sftp',
             'host' => env('SFTP_HOST'),
-            'port' => env('SFTP_PORT', 22),
+            'port' => (int) env('SFTP_PORT', 22),
             'username' => env('SFTP_USERNAME'),
             'password' => env('SFTP_PASSWORD'),
             'root' => env('SFTP_BASE_PATH', '/'),
